requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
dotenv
certifi
pymongo
Pyaml
dill

-e .